pyproject.toml
[tool.poetry]
name = "application-tracker"
version = "0.1.0"
description = ""
authors = ["Evgenii Gusev <ged1994@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
boto3 = "^1.34.7"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
django-storages = "^1.14.2"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.5"
gunicorn = "^21.2.0"
python-magic = "^0.4.27"
s3transfer = "^0.10.0"
django-split-settings = "^1.2.0"
psycopg2 = "^2.9.9"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]
colorlog = "^6.8.0"
pre-commit = "^3.6.0"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
